secure-slm-gateway-project/secure-slm-gateway-project/requirements.txt
datasets==2.20.0
huggingface-hub==0.24.7
sentence-transformers==2.7.0
transformers==4.40.2
tokenizers==0.19.1
xgboost==2.0.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
tqdm==4.67.1
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1
langgraph==0.0.40
